feat: support Registry
