+++ pkg/controller/util.go
@@ -1,0 +1,7 @@
+package controller
+
+import "os"
+
+func mkdirAll(p string) error {
+	return os.MkdirAll(p, 0o775) //nolint:gomnd,wrapcheck
+}
